make err() non-fatal
This plumbs an error-set and early returns all through qbe, and then
guards all sqbe api functions so that they can still be safely called in
error situations.

This seemed preferable to setjmp/longjmp even though it's more invasive
in qbe because there's very many places we'd have to setjmp, and it'd be
for every small instruction, etc.

An alternative approach would be to have blocks, instructions, etc.
build in API to an auxiliary structure, then setjmp once and call all
the qbe code to convert. See how flaky/error-prone the current version
is for a while first I think.